commit for order-service

+++ order-service/src/main/java/uz/nt/orderservice/scheduled/TimerTaskOrderedProducts.java
@@ -2,8 +2,10 @@ package uz.nt.orderservice.scheduled;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.EnableScheduling;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
+import shared.libs.dto.UserDto;
 import uz.nt.orderservice.client.ProductClient;
 import uz.nt.orderservice.dto.OrderedProductsDetail;
 import uz.nt.orderservice.entity.OrderedProductsRedis;
@@ -21,26 +23,29 @@ public class TimerTaskOrderedProducts {
     private final OrderedProductsRedisRepository redis;
     private final ProductClient productClient;
 
-    public void holdingTheOrderForFifteenMinutes(Integer orderId, Integer userId){
-
+    public void holdingTheOrderForFifteenMinutes(Integer orderId){
         Timer timer = new Timer();
         TimerTask timerTask = new TimerTask() {
             @Override
             public void run() {
-                Optional<Integer> optional = orderRepository.findByIdAndUserIdAndPayedIsFalse(orderId, userId);
-                if (optional.isPresent()){
-                    orderRepository.deleteById(orderId);
+                if(SecurityContextHolder.getContext().getAuthentication().getPrincipal() instanceof UserDto userDto) {
+                    Integer userId = userDto.getId();
+                    Optional<Integer> optional = orderRepository.findByIdAndUserIdAndPayedIsFalse(orderId, userId);
 
-                    Optional<OrderedProductsRedis> optionalRedis = redis.findById(orderId);
-                    if (optionalRedis.isPresent()) {
-                        OrderedProductsRedis orderedProductsRedis = optionalRedis.get();
-                        productClient.addProductAmountBackWard(orderedProductsRedis.getOrderedProductsList());
-                        redis.deleteById(orderId);
+                    if (optional.isPresent()) {
+                        orderRepository.deleteById(orderId);
+
+                        Optional<OrderedProductsRedis> optionalRedis = redis.findById(orderId);
+                        if (optionalRedis.isPresent()) {
+                            OrderedProductsRedis orderedProductsRedis = optionalRedis.get();
+                            productClient.addProductAmountBackWard(orderedProductsRedis.getOrderedProductsList());
+                            redis.deleteById(orderId);
+                        }
                     }
-                }
 
-                timer.cancel();
-                timer.purge();
+                    timer.cancel();
+                    timer.purge();
+                }
             }
         };
